+++ shulkr/git.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 import os.path
 from typing import Optional
 
-from git import Commit
+from git import Commit, Repo
 
 
 def get_blob(commit: Optional[Commit], path: str, repo_path: str) -> bytes:
@@ -19,3 +19,15 @@ def get_blob(commit: Optional[Commit], path: str, repo_path: str) -> bytes:
 		curr = curr[name]
 
 	return curr.data_stream.read().decode()
+
+
+def head_has_commits(repo: Repo) -> bool:
+	try:
+		repo.iter_commits()
+		return True
+
+	except ValueError as e:
+		if 'does not exist' in str(e):
+			return False
+
+		raise e