+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - `File exists` error when decompiling the second version
+- `--undo-renamed-vars` causing error
 
 ## [0.3.1] - 2021-12-30
 ### Changed